Removed old ICLR script and added workshops
